StringCondition -> StringFunction
